fix: stop removing from ES by gmcNumber when gmcNumber is null

+++ src/main/java/uk/nhs/hee/tis/revalidation/connection/service/ElasticSearchIndexUpdateHelper.java
@@ -29,7 +29,9 @@ public class ElasticSearchIndexUpdateHelper {
       elasticSearchService.addExceptionViews(getExceptionViews(connectionInfo));
     }
     else {
-      elasticSearchService.removeExceptionView(connectionInfo.getGmcReferenceNumber());
+      if (connectionInfo.getGmcReferenceNumber() != null) {
+        elasticSearchService.removeExceptionView(connectionInfo.getGmcReferenceNumber());
+      }
     }
   }
